requirements.txt
django==2.2
django-crispy-forms
djangorestframework
scikit-learn
pandas
numpy
Pillow
